pyproject.toml
[project]
name = "semeval-2026-task13"
version = "0.1.0"
description = "SemEval-2026 Task 13: Detecting Machine-Generated Code"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=1.1.0",
  "bitsandbytes>=0.44.0",
  "datasets>=3.2.0",
  "evaluate>=0.4.3",
  "loguru>=0.7.0",
  "numpy>=2.0.0",
  "pandas>=2.2.0",
  "peft>=0.13.0",
  "protobuf>=5.0.0",
  "pyarrow>=18.0.0",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.5.0",
  "sentencepiece>=0.2.0",
  "tiktoken>=0.12.0",
  "torch>=2.5.0",
  "transformers>=4.47.0",
]

[dependency-groups]
dev = [
  "ipykernel>=6.29.0",
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/semeval2026_task13"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
